backend/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
weasyprint==60.1
gunicorn==21.2.0
cryptography==41.0.7
